Add connect and disconnect functions to ESP32 network driver
To take advantage of network:wifi_scan/1 results the STA interface can be
enabled, but not immediately start a connection by adding the key `managed` to
the STA configuration when the driver is started.

Adds network:connect/1 to allow connecting to an access point after the driver
has been started in STA or STA+AP mode. If the function is used without
parameters a connection to the last configured access point will be started.

Adds network:disconnect/0 to disconnect a station from an access point.

The station mode disconnected callback now maintains the default behavior of
reconnecting to the last access point if the connection is lost, but if the
user defines a custom callback the automatic re-connection will not happen,
allowing for users to take advantage of scan results or some other means to
determine when and which access point to associate with.

Signed-off-by: Winford <winford@object.stream>