Add connection managment to ESP32 network driver
For more fine grained connection management in applications the driver
can now be started, without perfoming an inital connection by the use
of the key `managed` in the STA configuration.

Adds network:sta_connect/0,1 to allow connecting to an access point
after the driver has been started in STA or STA+AP mode. When used
without parameters a connection to the last configured access point
will be started.

Adds network:sta_disconnect/0 to disconnect a station from an access
point. The station mode disconnected default callback behavior is
unchanged unless a custom callback is provided, in which case the
automatic re-connection will not happen, allowing for users to take
advantage of scan results or some other means to determine when and
which access point to associate with.

The combination of the use of a disconnected callback and `managed`
mode allow for the use of `network:wifi_scan/0,1` (PR #1165), since the
wifi must not be connected to a station when perfoming a scan and old
implementation always started a connection immediatly and always
persisted in attempts to reconnect when disconnected.

Signed-off-by: Winford <winford@object.stream>