Handle wifi event 40 on ESP-IDF v5.2 builds
ESP-IDF version 5.2 introdued a new event (40), which indicates a home channel
change when devices are connected in station mode. These informational events
are not very useful for applications to act on, as this is expected when a
station connects to a new access point, or when the access point instructs
connected devices to change to a different channel. Rather than confuse
application developers with log messages for an unhandled event 40, by default
these events now do nothing, but if debug level logging is enabled in ESP-IDF
Kconfig setting an informative debug log entry will be printed on the console.

Signed-off-by: Winford <winford@object.stream>