Fix test_gc test with BEAM
`test_gc` was disabled because it failed more or less randomly with OTP26.
It seems to be related to a change in how the eval string from the command line
is processed by OTP, allocating more memory and therefore making the effect of
garbage collector less obvious. Fix it by spawning a fresh process for the
test.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>