Adjust `Doxygen.in` config to reduce build warnings and fix rendering errors
Changes to the `Doxyfile.in` configuration file
- `MACRO_EXPANSION` and `EXTRACT_STATIC` set to `YES` allow including documented static functions
as well as expanding macros so Doxygen can correctly process some functions, this greatly reduces
the number of extraneous warning emitted by Doxygen during processing.
- Adds several definitions to `PREDEFINED` so these conditionally included functions will be found
and included in the documentation.
- `INTERACTIVE_SVG = YES` will improve how svg are displayed by modern browsers. This was set to off
by default to accommodate old versions if Internet Explorer that have been deprecated for many years.
- `INLINE_SOURCES` allows the values of some enums and class member values to be displayed.
- `HIDE_IN_BODY_DOCS` will prevent adding any comments within functions that are documented to the
published documentation. Any of these comments in our code a certainly meant for internal notation and
not expected to be published in the on-line documentation.
- `SHOW_NAMESPACES` has been changed to `NO`. Several pseudo-namespaces have been used to reduce
warnings during the documentation build process about duplicate declarations for function and data
structures that are included in more than one place in the documentation. These namespaces do not need
to be shown or pollute links with the extra tags.
- Enable `WARN_NO_PARAMDOC` so helpful warning will be emitted when documented functions are missing
parameters in the documentation.
- Add `defaultatoms.c`, `scheduler.c`, and `tempstack.h` to the files excluded form Doxygen processing,
these files were responsible for a large number of warning and add nothing to the documentation. These
files are also added to a list of omitted files along with `opcodesswicth.h`, which was previosly added
to the Doxyfile.in `EXCLUDE_PATERNS`, that CMakeLists.txt will remove the skeleton .rst files for, since
there will be no `doxygenfile` data to insert into the placeholders.

Signed-off-by: Winford <winford@object.stream>