Fix Sphinx namespace collisions to reduce doc build duplicate definition warnings
This reduces unnecessary warnings that occur when a function, structure, or
enumeration is included in more than one place in the documentation. To allow
the display of collected function and data structures as well as individual
source files without unnecessary warnings each source now gets the source file
name appended to the libAtomVM namespace used for documenting the C APIs.

Signed-off-by: Winford <winford@object.stream>