Enhance the appearance of the documentation "versions" menu
This makes some minor adjustments to appearance of the downloads and alternate
versions sub-menu at the bottom of the sidebar. A download icon will appear next
to the name of the currently displayed version to help draw attention to the
availablity of downloadable versions of the documentation, this was brought
up on the [erlangforums](https://erlangforums.com/t/atomvm-0-6-3-released/3852).
The whole word "version:" is used, and moved ouside the highlighted name, rather
than the "v:" that was previously shown next to the name with the same
highlighting, this had the effect of making the menu easy to overlook, and the
purpose was not clear. Icons have been added next to the "Versions" and
"Downloads" sub-heading names in the opened versions menu to enhance the overall
appearance.

Signed-off-by: Winford <winford@object.stream>