Publish documentation for branches and tags
These changes will allow publishing documentation for branches (in addition to
`main`, which is already published), while continuing to support publishing for
release tags. The menu listing for `main` has been enhanced to show the link as
`main branch (unstable)`, other branch names will appear as (unreleased), like:
`release-0.6 branch (unreleased)`. The appearance of links to (tagged) released
versions of the documentation are unchanged. This should make it easy for users
to identify the correct version of the documentation, and not mistake
`release-0.6` for a stable 0.6 release version of the documentation, and easily
identify the docs version that matches a branch they are interested in working
on, by differentiating branch versions of the documentation.

Signed-off-by: Winford <winford@object.stream>