Fix publish docs for branches
There was a bug in the tests for GitHub environment variables that led to the
documentation build system using an undefined environment variable which
resulted in an empty branch name. Determining the correct branch or tag name
from within cmake using GitHub environment variables is unreliable, so instead
the correct name is determined by the workflows, which set an environment
variable that can be used by cmake.

Signed-off-by: Winford <winford@object.stream>