Correction to publish_docs workflow for branches
The github `actions/checkout@v4` only tracks the current branch, even when
`fetch-depth: 0` is used, this causes branch names other than the current one
to be missed when creating the html navigation menus. This is solved by adding
a step to track all of the remote branches that were fetched during checkout.

Signed-off-by: Winford <winford@object.stream>