Fix bogus out_of_memory in `unicode:characters_to_list`
Quoting `man malloc`:
"If size is 0, then malloc() returns either NULL, or a unique pointer
value that can later be successfully passed to free()."

Platforms such as esp-idf prefers NULL in that case.

So `IS_NULL_PTR` may fail in situations such as when dealing with `malloc(size)`
when size is 0 (due to a 0 length binary, leading to out of memory when
doing `unicode:characters_to_list(<<>>)` on ESP32 but not on other
platforms.

Signed-off-by: Davide Bettio <davide@uninstall.it>