Check for content change before publishing
Adds a check to make sure there are acutally changes to be commited and pushed.
This is needed to address a situation where the publish docs workflow is
triggered, but there are no changes to the documentation content.

Changes `cancel-in-progress` to false, to avoid missing publishing documentation,
when another publish workflow is triggerd by a commit that might not have any
changes to the documentation to add.

Signed-off-by: Winford <winford@object.stream>