Add `atomvm:subprocess/4` performing pipe/fork/execve`
This is implemented using pipe/fork/execve or posix_spawn depending on
platforms and availability of `HAVE_POSIX_SPAWN_CLOEXEC_DEFAULT` attribute.
The VM does not leak any file descriptor to the subprocess except for the
pipe used to read stdout.

The function is explicitely disabled on microcontrollers and is meant to be
used on generic_unix platform, for example for tests.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>