Fix ESP32 network:stop/0
Removes all of the registered event handlers when the driver is stopped.

No longer deletes default_event_loop when the driver is stopped. Deleting this loop was a mistake
since other processes may use the default loop and it is normally created at boot up, not as part
of the network initialization.

Adds a network:stop/0 and a second start to the test_wifi_example esp32-simtest to help catch
future problems with stopping and restarting the network.

Closes #1449

Signed-off-by: Winford <winford@object.stream>