Safer creation of atoms in ESP32 network_driver.c
Removes the up-fron declaration of `static const char *const` atom strings in favor of inline
ATOM_STR() for better redability. Changes to use of `globalcontext_existing_term_from_atom_string`
where possible. AP and STA mode specific atoms are only created for the interface(s) that are used.
All atoms that are created by the driver are now checked to be certain the creation was sucessful.

Addresses concerns raised in issue #1442 for the ESP32 network driver.

Signed-off-by: Winford <winford@object.stream>