Add test_gpio.erl to ESP32 test suite
Adds a test to run on hardware with a jumper wire, or in the wokwi simulator to test the basic
esp32 gpio driver funtionality and error handling, with tests for all boards when full_sim_test
is run.

The sim-test for P4 depends on PR #1438 being merged to be able to select the correct pins for the
device for the the GPIO tests.

Signed-off-by: Winford <winford@object.stream>