Fix publish-docs to keep artifacts out of the committed html
There were .doctree cache artifacts being included in the 'doc/html', 'doc/html/pdf', and
'doc/html/epub' folders that are only used during Sphinx generation of the content. Moves the
location of these files to a common location in the build/doc folder outside of the html folder
to be committed. This should have the added benefit of reducing the workload, since all three
versions use can share the same cache files.

Signed-off-by: Winford <winford@object.stream>