Remove pre-release and old versions from the online documentation
Removed the v0.5.0 and pre-release versions of v0.6.0 from the menu in the published documetation.
This will direct users to more recent stable relese versions, and help keep the size of the
"Production" branch of atomvm_www below the 1 GB limit for GitHub Pages branches.

Note: the directories for these old version will be removed from the atomvm_www Production branch
in a PR in that repo, these changes just prvent them from being incuded in the navigation menu.

Signed-off-by: Winford <winford@object.stream>