Fix refcount calculation when copying data to heaps
Ref count should be incremented whenever the ref counted binary is added to
the mso list and decremented when the mso list is swept.

Also refc_binaries are allocated with a ref count of 0 until they are made
a term, with the exception of enif_alloc_resource that should be balanced
by an enif_release_resource following OTP semantic.

Also add a new test for heap operations including gc.

Also fix race conditions with demonitor in both otp_socket and emscripten
code that yielded a crash by over decrementing ref count.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>