Cleanup ledc driver
Removes an include for a file that does not exist. It was dropped during the PR introducing the ledc
driver, but the include line was left, as well as several uses of macros defined in the file.

Removes undefined macros, replacing them with their values, and documents the meaning of the values.

Signed-off-by: Winford <winford@object.stream>