Improve error logging in ESP32 i2c driver
Logged errors now include the decoded error string, rather than printing the error number, which
must be looked up in the ESP-IDF documentation for the function returning it. Especially for
application developers unfamiliar with ESP-IDF it may not even be obvious which IDF function was
called that returned the error when a erlang i2c function is used.

Signed-off-by: Winford <winford@object.stream>