Fix a concurrency issue yielding a timeout bug
Ensure that `scheduler_make_ready` and `scheduler_set_timeout` calls
`sys_signal`. As `global->waiting_scheduler` was cached or read without
holding `global->schedulers_mutex`, in some cases `sys_signal` was not
called yielding a timeout bug where a message is in the queue but the process
is not awaken.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>