Fix 0x2b error in `nif_prim_file_get_cwd_0` (#115)
Fixes EXW-197

The problem was caused by allocating a tuple without initializing it.
This meant the 2 words had GCed item (containing moved marker `0x2B`) or
garbage. When calling `memory_ensure_free_with_roots`, a shallow copy of
that tuple (it was the root) attempted to move that garbage and was
failing.

Signed-off-by: Mateusz Front <mateusz.front@swmansion.com>