Add bigint implementation (`intn.c`)
`intn.c` contains functions for manipulating bigints (array of n
digits):
- `intn_mulmns`, `intn_divmnu`, and `nlz` are from Hacker's Delight
- Other functions such as `intn_addmns` are original work

This version is an attempt with numbers in 2nd complement, so division
required a wrapper for calling `divmnu` using an absolute value (this
specific function has not been broadly tested yet).

Given functions are limited to a maximum size for inputs and outputs
that is defined in `INTN_MAX_IN_LEN` and `INTN_MAX_RES_LEN`.
That's the reason it is called intn and not bigint.

Signed-off-by: Davide Bettio <davide@uninstall.it>