Use sign bit for big integers (instead of 2-complement)
Some operations in 2-complement turns to be quite complex, they
require more code, and also more stack space for storing abs value.
Hence using a dedicated sign bit (as Erlang does) turns to be an easier and
pragmatic approach.

This approach makes possible having sign bit outside the numeric
payload, so the supported range is -(2^256 - 1)..+(2^256 - 1).
They might be called int257, but it would be quite confusing.

Sign bit is stored in boxed header, outside of the numeric payload.

Signed-off-by: Davide Bettio <davide@uninstall.it>