Add wifi support via esp-wifi-external for ESP32P4
Adds components/avm_builtins/idf_component.yml that conditionally retreivies and includes the
espressif/esp-wifi-external component for when building for the `esp32p4` target.

Chip specific sdkconfig.defaults.esp32p4 enable the on-board PSRAM, sets the flash size to 16M
(since this is the standard ammount incuded in P4 chips), and configure the use on the on-board
ESP32C6 WiFi module as the wifi host device.

Adds ESP32P4 build tests to CI workflow.

Signed-off-by: Winford <winford@object.stream>