module: do not abort when OOM while loading literal
Literals are encoded as external terms, so they might make use
of heap fragments and their memory allocation might fail.
Existing code already deal with this, so we can just raise instead of
abort().

There is still a very very low change we found a corrupt external term,
but we are not able to distinguish the situations, so we also print a
warning.

Signed-off-by: Davide Bettio <davide@uninstall.it>