opcodesswitch.h: fix possible memory corruption with boxed imm integers
`memory_ensure_free_opt` does not guarantee that a fragment is always
created, when space is available, allocation happens on current heap:
```
153     if (alloc_mode == MEMORY_NO_GC) {
154         if (free_space < size) {
```
In some specific circumstances this might lead to a memory corruption.

Checked fact:
Q: Does `allocate_heap` opcode (on its `HeapNeed` parameter) accounts also for
the number of words required for boxed integers used for storing immediate
values on the heap?
A: No.
Note: the compiler doesn't assume anything about the size of integers, because
the same BEAM code must work on both 32-bit and 64-bit systems.

Signed-off-by: Davide Bettio <davide@uninstall.it>