opcodesswitch: fix false positive CodeQL warnings
Cast terms to uintptr_t and intptr_t when an integer is required, to
avoid "Passing a term to a function expecting a non-term" and "Passing a
non-term to a function expecting a term" warnings.

Also change Context.cp type to term for the same reason.

This change is 100% safe since term is already defined as uintptr_t.

Signed-off-by: Davide Bettio <davide@uninstall.it>