Start with `init:boot/1`
If init module is found, `init:boot/1` is called instead of the startup module.
This function takes a list of parameters, towards a CLI interface.

The logic is factorized across platforms through a new `globalcontext_run`
function.

Because some platforms allowed 0 for success, start/0 returning ok or 0 is
now considered a success across platforms.

Also fix documentation related to this return value and its usage.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>