Serialize and unserialize functions
Support functions with `term_to_binary/1` and `binary_to_term/1`.

MD5 sum of the module is not computed and module version matching is based on `old_uniq` field which is deprecated and likely to be removed in OTP29. If a function is serialized by AtomVM, the MD5 sum is replaced by zeroes, thus BEAM may not recognize it. If a function is serialized by BEAM it can be unserialized and executed by AtomVM if the same beam is on the AtomVM node.

If a function is unserialized on AtomVM but the module doesn’t exist, it will always raise `undef` when executed, even if the module is later loaded. This limitation will be fixed once AtomVM will support code reloading.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>