Align gen_tcp inet backend error returns with OTP returns
Fix several discrepancies with OTP error returns from gen_tcp:connect/3 in the gen_tcp_inet.erl
backend. Connecting to an unresolvable hostname now returns `{error, nxdomain}` which was
previously `{error, {getaddrinfo, Err}}` where the value of Err varies by platform and host OS the
VM was compiled on (for some Linux systems this value was `-2`, and on FreeBSD the value was `8`).
Connecting with an invalid address such as an integer, incorrect size of address tuple, or out of
ip address range tuple will raise a `badarg`, the old error returns varied depending on what
address was sent to the driver, none of which were the correct behavior of raising `badarg`.

Adds tests that would have previously failed to ensure error returns are consistent with OTP.

Signed-off-by: Winford <winford@object.stream>