Fix crash when firing resource monitor, typically for distribution
There was a race condition where a context being monitored by a resource is
being terminated at the same time as the resource. The resource message to
unmonitor would be lost by the process and the process tried nevertheless
to fire the monitor, causing a crash (worst case) or a valgrind error
(best case).

Signed-off-by: Paul Guyot <pguyot@kallisys.net>