Enable/Make AVM_PRINT_PROCESS_CRASH_DUMPS work on embedded platforms
AVM_PRINT_PROCESS_CRASH_DUMPS was introduced with bcfdcc69d (PR #1799),
but the option hasn't been exposed / set on embedded targets (such as
esp32, stm32 and rp2).

The reason is that esp32 CMakeLists.txt doesn't include the root
CMakeLists.txt, and the libAtomVM one doesn't have that option defined
(it is defined on the root CMakeLists.txt), so that option didn't have a
default value.

Signed-off-by: Davide Bettio <davide@uninstall.it>